[refs #37] As of astroid 1.3.3, class method resolution order is taken into consideration: https://bitbucket.org/logilab/astroid/commits/b1d8107715a4ceeeeb3f4a9065b67817e8cdc643
This caused an infinite loop when using pylint-django, as the model and form field transforms extended the base Django object then replaced it. This caused the MRO lookup to find a class as its own ancestor.

The model and form field transforms have been simplified, so that instead of building a module and replacing each class individually, instead, the correct "dummy" parent class is simply appended. This means that the original parent node (eg, CharField) is not replaced by itself. This also has the side effect of no longer confusing pylint when it comes to __init__/__new__ differences for field classes which extend "immutable" base classes as a result of this transform.