Suppress no-member errors for RelatedManager fields
seen in a real project where we have `product.build.first()` where
Build.product is a FK to Product with related_name=build all of
which causes factory to build a RelatedManager on the fly.

For some reason I can't reproduce with the existing test but when
I make this change it resolves the problem in the original code base
that I was inspecting.