Pin Faker to Prevent Asteroid Crash
There's some sort of issue with version 5.0.2 of Faker which updates a
lot of code to use the pathlib library instead of os utils. This causes
the sys.path_importer_cache to get filled with PosixPath objects instead
of strings as it's keys, which causes an asteroid crash.

For now, we can just pin Faker to a version that still works until the
upstream issue is resolved.